First commit
So this was a test to see how useful AI can be in generating content. Text and images were AI generated, and I threw them into a simple Nuxt page. Probably not a great curriculum for now, and I honestly haven't really read much of it. But it's a start.