Revise "For educators" section in the intro page
Changed the text to be more about our ambition that the guide is used in educational contexts. We have not actually tried it out, or done anything to make it great for education. So now the text is less a statement, than an ambition, as well as an invite to contributions.