Chapter 1 revision based on reading session (#14)
* Chapter 1 revision based on reading session

There was quite a bit of feedback from the reading session (https://github.com/BitcoinDesign/Meta/issues/712). I tried to implement as much of it as possible in this first commit, but I'll have to come back it with fresh eyes next week and iterate further.

* Re-reading and revising

Coming back with a fresh brain. Caught some typos and rewrote some sentences to flow better.