Revision of chapter 2 based on reading session
Trying to address some of the feedback we came up with. I have to come back to this one and iterate a little further.

I stayed away from the success stories, as I think this is a separate and larger mini-project to tackle. It requires research and digging up those great examples and goes beyond copy revision.

This also introduces a new type of link - a chapter link. It has a distinct look from external links, so readers can easily identify internal cross-references. Curious to hear what others think of it.