Chapter 7 revision based on reading session
This turned into a much bigger rewrite, including a change of the page name. In our reading session, we did not find the framing of contributing assets very clear. This updated version essentially speaks to publicly documenting your design process and work. Goal is to for design to be transparent, so others can easily catch and keep up and contribute.

The page still feels like a little bit of a patchwork, but I think I need to come back to it and/or feedback from others. So I'm putting this PR out there now for conversation. So what do you think?