Chapter 5 revision based on reading session (#24)
* Chapter 5 revision based on reading session

There are quite a few changes to this page. I tried a different approach this time around. Instead of manually implementing the feedback we had, I gave it to Claude again and asked it to do a rewrite. I quite liked what I got back. So I took the rewrite and the original and edited manually section by section, while also referencing our feedback notes.

So it might a little more tricky to review these changes, since it's not just minor tweaks. I'd recommend having the current page and the preview page of this PR open next to each other, and review one chapter at a time on both versions.

* Adding in additions from Eriol

Two changes based on these comments:
- https://github.com/GBKS/opendesign.guide/pull/24#discussion_r1846441895
- https://github.com/GBKS/opendesign.guide/pull/24#discussion_r1846528511