Chapter 8 revision based on reading session
Another page where we decided in the reading session to do a bigger rewrite. The original content repeated a lot of things already discussed in earlier chapters, and it was (as usual) short on examples. The new focus with this rewrite here is to give more specific examples on how to contribute as a designer, along with examples and reference link.

While working on this, I ended up with a pretty good paragraph on user groups. That was not a good fit for this page in the end, so I moved it to the design process page.

I don't think this initial commit is super good, but I'd like to get it out there for feedback. Please let me know what you think.