Chapter 7 revision based on reading session (#26)
* Chapter 7 revision based on reading session

This turned into a much bigger rewrite, including a change of the page name. In our reading session, we did not find the framing of contributing assets very clear. This updated version essentially speaks to publicly documenting your design process and work. Goal is to for design to be transparent, so others can easily catch and keep up and contribute.

The page still feels like a little bit of a patchwork, but I think I need to come back to it and/or feedback from others. So I'm putting this PR out there now for conversation. So what do you think?

* Apply suggestions from code review

Co-authored-by: paperpsych <149538435+paperpsych@users.noreply.github.com>

* Trying to fix the PR

Can't commit any proposed changes by Paperpsych on Github, and Github is stuck on checking whether merging can be done automatically. Maybe another commit will snap it out of it's problems.

* Fix formatting

* Renaming the page

---------

Co-authored-by: paperpsych <149538435+paperpsych@users.noreply.github.com>