Chapter 12 (final project) revision based on reading session
This is a light revision of chapter 12. The biggest change is the new page title. In the reading session, we felt that "first project" is more appropriate than "final project". There is no accreditation that readers receive from completing this. It's more like they are starting a new journey with this project, where they go from learners to doers.