Chapter 10 (Advanced Topics) revision based on reading session (#30)
* Chapter 10 (Advanced Topics) revision based on reading session

This is more of a light revision with text tweaks and formatting changes. Hope I managed to get all the feedback in we had in the reading session.

* Apply suggestions from code review

Co-authored-by: paperpsych <149538435+paperpsych@users.noreply.github.com>

* Minor tweaks and fixes

* CSS tweak

Trying to get a little more spacing between list items in the article main content. They are very condensed. Trick is to make it look good for short bullet lists as well as ones with multi-line items.

---------

Co-authored-by: paperpsych <149538435+paperpsych@users.noreply.github.com>